Removed uneccessary dependency and moved the toggles to a new file
