+++ package.json
@@ -21,7 +21,6 @@
     "webpack-cli": "^3.3.6"
   },
   "dependencies": {
-    "@storybook/react": "^5.2.1",
     "axios": "^0.19.0",
     "body-parser": "^1.19.0",
     "css-loader": "^3.2.0",
